backend/pyproject.toml
[tool.poetry]
name = "github-rag"
version = "0.1.0"
description = ""
authors = ["Johannes Hayer <contact@jhayer.tech>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.31.0"
openai = "^1.45.0"  # Updated to latest version
psycopg2-binary = "^2.9.6"
python-dotenv = "^1.0.1"
numpy = "^2.1.1"
termcolor = "^2.4.0"
fastapi = "^0.100.0"
uvicorn = "^0.23.0"
pydantic = "^2.1.1"
litellm = "^1.46.6"  # Added LiteLLM
sse-starlette = "^2.1.3"
websockets = "^13.0.1"


[tool.poetry.dev-dependencies]
pytest = "^7.4.0"
httpx = "^0.24.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "uvicorn server:app --reload"